Set bearing by using geolocate fitBoundsOptions
I'm trying to get the camera to point in the direction of travel (when
the user's heading is available through the navigator.geoLocation API at
least), but if you set the map bearing using map.setBearing() then
Mapbox GL JS will immediately stop tracking the user.

This alternative method sets the bearing in the GeolocateControl's
fitBoundsOptions. I've only tested on a desktop (faking heading changes
rather than getting it from the brower), but it seems like it might do
the trick on mobile.